logs: Indicate current day by keeping panel heading sticky
Currently the current day was shown in the top filter.
This dropdown had values 'Recent', 'Current Boot'... but when some data
were loaded it suddenly showed some date. Or sometimes 'Go To'. When it
showed some date, it did not mean it showed just items from that given
day, but that you are currently scrolled over that day.
Dropdown that has value that you cannot actually select is rather
confusing. Also it does not indicate what value you used to filter, as
do any other filters right next to it.

Now the panel heading that contains the date is stuck underneath the
filter panel and therefore it is obvious on which day you currently
without obscuring the filters.