logs: Improve mobile mode
The main issue was that the filters were stacked one on top of each
other and it could take significant amount of the screen. Also I plan to
add additional filters, so we could be looking at almost half of the
screen taken by just filters. And it was almost impossible to have the
date sticky panel, as we never knew, how much the filters take.

So when the filters wrap to another line, all but the first one are
hidden and 'Show all filters' button is displayed. This button shows all
additional filters. It is also possible to hide them back in.

Also labels and inputs were not properly kept together, so it often
happened that label would be on one line left to previous input and
the input would be on the next line. That was confusing. Now label is
always on the left, or on separate line above the input (on very narrow
screens).